refactor: Simplify inheritance
