+++ gym_simplifiedtetris/envs/simplified_tetris_binary_env.py
@@ -5,21 +5,13 @@ from gym import spaces
 
 from ..register import register
 from .simplified_tetris_base_env import SimplifiedTetrisBaseEnv
-from .simplified_tetris_engine import SimplifiedTetrisEngine
 
 
 class SimplifiedTetrisBinaryEnv(SimplifiedTetrisBaseEnv):
     """
     A class representing a custom Gym env for Tetris, where the observation space
     is the binary representation of the grid plus the current piece's id.
-
-    :param grid_dims: the grid's dimensions.
-    :param piece_size: the size of the pieces in use.
     """
-
-    @property
-    def action_space(self) -> spaces.Discrete:
-        return spaces.Discrete(self._num_actions_)
 
     @property
     def observation_space(self) -> spaces.Box:
@@ -31,87 +23,9 @@ class SimplifiedTetrisBinaryEnv(SimplifiedTetrisBaseEnv):
             dtype=np.int,
         )
 
-    def __init__(self, grid_dims: Sequence[int], piece_size: int):
-        super(SimplifiedTetrisBinaryEnv, self).__init__(
-            grid_dims=grid_dims, piece_size=piece_size
-        )
-
-        self._engine = SimplifiedTetrisEngine(
-            grid_dims=grid_dims,
-            piece_size=piece_size,
-            num_pieces=self._num_pieces_,
-            num_actions=self._num_actions_,
-        )
-
-    def __str__(self) -> str:
-        return np.array(self._engine._grid.T, dtype=int).__str__()
-
-    def __repr__(self) -> str:
-        return f"{self.__class__.__name__}(({self._height_!r}, {self._width_!r}), {self._piece_size_!r})"
-
-    def reset(self) -> np.array:
-        self._engine._reset()
-        return self._get_obs()
-
-    def step(self, action: int) -> Tuple[np.array, float, bool, Dict[str, Any]]:
-        """
-        Hard drops the current piece according to the argument provided. Terminates
-        the game if a condition is met. Otherwise, a new piece is selected, and the
-        anchor is reset.
-
-        :param action: the action to be taken.
-        :return: the next observation, reward, game termination indicator, and env info.
-        """
-        info = {}
-
-        # Get the translation and rotation.
-        translation, self._engine._rotation = self._engine._all_available_actions[
-            self._get_obs()[-1]
-        ][action]
-
-        # Set the anchor.
-        self._engine._anchor = [translation, self._piece_size_ - 1]
-
-        # Hard drop the piece and update the grid.
-        self._engine._hard_drop()
-        self._engine._update_grid(True)
-
-        # Game terminates if any of the dropped piece's blocks occupies any of the
-        # top piece_size rows, before any full rows are cleared.
-        if np.any(self._engine._grid[:, : self._piece_size_]):
-            info["num_rows_cleared"] = 0
-            self._engine._final_scores = np.append(
-                self._engine._final_scores, self._engine._score
-            )
-            return self._get_obs(), self._get_terminal_reward(), True, info
-
-        # Get the reward and update the score.
-        reward, num_rows_cleared = self._get_reward()
-        self._engine._score += num_rows_cleared
-
-        # Get a new piece and update the anchor.
-        self._engine._update_coords_and_anchor()
-
-        # Update the info.
-        info["num_rows_cleared"] = num_rows_cleared
-
-        return self._get_obs(), float(reward), False, info
-
-    def render(self, mode: Optional[str] = "human") -> np.ndarray:
-        return self._engine._render(mode)
-
-    def close(self) -> None:
-        return self._engine._close()
-
     def _get_obs(self) -> np.array:
         current_grid = self._engine._grid.flatten()
         return np.append(current_grid, self._engine._current_piece_id)
-
-    def _get_reward(self) -> Tuple[float, int]:
-        return self._engine._get_reward()
-
-    def _get_terminal_reward(self) -> float:
-        return 0.0
 
 
 register(
